Server support
* Added Acceptor to accept connections in server mode
* No more leaks detected with 100 connections in parallel to server
* No leaks between client and server (only connection process tested) !
* No more detected SIGSEGV / ABORT !
* Still data processing to test on server side, and let client / server communicates together

+++ include/aio4c/types.h
@@ -57,8 +57,8 @@ typedef struct pollfd aio4c_poll_t;
 
 typedef int aio4c_pipe_t[2];
 
-#define AIO4C_PIPE_READ 0
+#define AIO4C_PIPE_READ (short)0
 
-#define AIO4C_PIPE_WRITE 1
+#define AIO4C_PIPE_WRITE (short)1
 
 #endif
